refactor: decompose PDF quality responsibilities
